backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
Faker==37.3.0
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
